chore: improve the embed guide
