+++ src/components/pages/embed/PreviewVariantsShowcase.js
@@ -1,0 +1,108 @@
+import React, { useState, useEffect } from 'react'
+import styled from 'styled-components'
+import { theme } from 'theme'
+
+import Box from 'components/elements/Box'
+
+import {
+  HeroCard,
+  OneLineCard,
+  TelegramCard,
+  TwitterCard,
+  NotificationCard,
+  ChatBubbleCard,
+  TweetCard,
+  STRIPE_DEMO_DATA
+} from './PreviewCards'
+
+const PREVIEW_URL = 'https://stripe.com'
+const PREVIEW_INTERVAL_MS = 3000
+const PREVIEW_FADE_MS = 560
+
+const PREVIEW_VARIANTS = [
+  { component: HeroCard, name: 'Standard Rich Card Preview' },
+  { component: OneLineCard, name: 'Compact Inline Preview' },
+  { component: TelegramCard, name: 'Telegram Style Link Card' },
+  { component: TwitterCard, name: 'Twitter / X Summary Card' },
+  { component: NotificationCard, name: 'iOS Style Notification Card' },
+  { component: ChatBubbleCard, name: 'WhatsApp Style Chat Bubble' },
+  { component: TweetCard, name: 'Embedded Tweet Widget' }
+]
+
+const PreviewStage = styled(Box)`
+  position: relative;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  ${theme({ minHeight: ['320px', '360px', '400px', '400px'] })};
+`
+
+const PreviewLayer = styled(Box)`
+  position: absolute;
+  inset: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 8px;
+  transition: opacity ${PREVIEW_FADE_MS}ms cubic-bezier(0.22, 1, 0.36, 1),
+    transform ${PREVIEW_FADE_MS}ms cubic-bezier(0.22, 1, 0.36, 1),
+    filter ${PREVIEW_FADE_MS}ms cubic-bezier(0.22, 1, 0.36, 1);
+  opacity: ${({ $active }) => ($active ? 1 : 0)};
+  transform: ${({ $active }) => ($active ? 'scale(1)' : 'scale(0.96)')};
+  filter: ${({ $active }) => ($active ? 'blur(0)' : 'blur(6px)')};
+  pointer-events: ${({ $active }) => ($active ? 'auto' : 'none')};
+
+  @media (prefers-reduced-motion: reduce) {
+    transform: none;
+    filter: none;
+    transition: opacity ${PREVIEW_FADE_MS}ms ease;
+  }
+`
+
+export const PreviewVariantsShowcase = () => {
+  const [data, setData] = useState(STRIPE_DEMO_DATA)
+  const [index, setIndex] = useState(0)
+
+  useEffect(() => {
+    let cancelled = false
+    fetch(`https://api.microlink.io/?url=${encodeURIComponent(PREVIEW_URL)}`)
+      .then(r => r.json())
+      .then(json => {
+        if (cancelled) return
+        if (json?.status === 'success' && json.data) setData(json.data)
+      })
+      .catch(() => {})
+    return () => {
+      cancelled = true
+    }
+  }, [])
+
+  useEffect(() => {
+    const id = setInterval(() => {
+      setIndex(i => (i + 1) % PREVIEW_VARIANTS.length)
+    }, PREVIEW_INTERVAL_MS)
+    return () => clearInterval(id)
+  }, [])
+
+  return (
+    <PreviewStage>
+      {PREVIEW_VARIANTS.map(({ component: Variant, name }, i) => (
+        <PreviewLayer key={i} $active={i === index} aria-hidden={i !== index}>
+          <h3
+            style={{
+              position: 'absolute',
+              width: 1,
+              height: 1,
+              overflow: 'hidden',
+              clip: 'rect(0,0,0,0)'
+            }}
+          >
+            {name}
+          </h3>
+          <Variant data={data} />
+        </PreviewLayer>
+      ))}
+    </PreviewStage>
+  )
+}
