+++ src/pages/alternative/iframely.js
@@ -29,7 +29,8 @@ import { Link } from 'components/elements/Link'
 import Meta from 'components/elements/Meta/Meta'
 import SubheadBase from 'components/elements/Subhead'
 import Text from 'components/elements/Text'
-import { InteractiveExample, PreviewVariantsShowcase } from 'pages/embed'
+import { PreviewVariantsShowcase } from 'components/pages/embed'
+import { InteractiveExample } from 'pages/embed'
 
 const Heading = withTitle(HeadingBase)
 const Subhead = withTitle(SubheadBase)
@@ -839,7 +840,8 @@ const Hero = () => (
         <ArrowLink
           href='/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'iframely' })}
+            trackEvent('alternative cta', { competitor: 'iframely' })
+          }
         >
           Get Started Free
         </ArrowLink>
@@ -1216,7 +1218,8 @@ const CTASection = () => (
         <ArrowLink
           href='/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'iframely' })}
+            trackEvent('alternative cta', { competitor: 'iframely' })
+          }
           css={theme({ fontSize: 3, px: 5, py: 3, color: 'white' })}
         >
           Start Building Free
@@ -1395,7 +1398,8 @@ const TryItSection = () => (
         <Link
           href='/docs/guides/embed'
           onClick={() =>
-            trackEvent('alternative cta', { competitor: 'iframely' })}
+            trackEvent('alternative cta', { competitor: 'iframely' })
+          }
           css={theme({ fontSize: CTA_LINK_FONT_SIZE })}
         >
           Start now for free
